requirements.txt
google-generativeai==0.8.4
python-dotenv==1.1.0
pytest==8.3.5
pytest-mock==3.14.0
PySide6==6.8.3
Pillow==11.1.0
mcp==0.1.0
anthropic==0.19.0 